+++ .gitignore
@@ -1,7 +1,5 @@
 ##########################
 # Dependency Directories #
 ##########################
-/client/node_modules
-/server/node_modules
-/client/.env
-/server/.env
+node_modules
+.env
