+++ src/App.js
@@ -1,0 +1,22 @@
+import React, { useState, useEffect } from "react"
+
+import './App.css';
+import Routing from "./routes/index";
+
+function App() {
+
+
+
+ 
+
+  return (
+    <div className="App">
+      <header className="App-header">
+
+      </header>
+      <Routing />
+    </div>
+  );
+}
+
+export default App;
